Handle case where is not possible to open the display
This is a corner case where I'm running awesome inside a VirtualBox VM
for which the display is not really working perfectly.

```
1 manzo@archlinux ~ % xbacklight -get                                                                                               :(
RANDR Query Version returned error -1
1 manzo@archlinux ~ % xrandr --query                                                                                                :(
Can't open display
```

I hope this may be useful in other cases because an error here has
prevented all my configs to be loaded.